+++ lib/factories/create_transcriber_repository.dart
@@ -7,11 +7,11 @@ import '../repositories/transcriber_repository.dart';
 import 'http_client.dart';
 
 TranscribeModel createTranscriberRepository() {
-  var fac = DynamicFactories.speeachToText;
+  var fac = DynamicFactories.transcriber;
   if (fac != null) return fac();
   return TranscriberRepository(
     dio: httpClient,
     model: AudioModel.whisper1,
-    language: VitGptConfiguration.transcriptionLanguage,
+    language: VitGptDartConfiguration.transcriptionLanguage,
   );
 }