+++ lib/usecases/local_storage/audio/save_transcription_language.dart
@@ -4,5 +4,5 @@ import 'package:vit_gpt_dart_api/data/dynamic_factories.dart';
 Future<void> saveTranscriptionLanguage(String lang) async {
   var rep = DynamicFactories.localStorage;
   await rep.saveTranscriptionLanguage(lang);
-  VitGptConfiguration.transcriptionLanguage = lang;
+  VitGptDartConfiguration.transcriptionLanguage = lang;
 }