+++ lib/usecases/local_storage/audio/save_tts_quality.dart
@@ -4,5 +4,5 @@ import 'package:vit_gpt_dart_api/data/dynamic_factories.dart';
 Future<void> saveTtsQuality(bool highQuality) async {
   var rep = DynamicFactories.localStorage;
   await rep.saveTtsQuality(highQuality);
-  VitGptConfiguration.useHighQualityTts = highQuality;
+  VitGptDartConfiguration.useHighQualityTts = highQuality;
 }